requirements.txt
alembic==1.15.1
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
certifi==2025.1.31
cffi==1.17.1
chardet==3.0.4
click==8.1.8
cryptography==44.0.2
dnspython==2.7.0
docutils==0.21.2
ecdsa==0.19.1
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.11
fastapi-jwt-auth==0.5.0
filelock==3.0.12
h11==0.14.0
httptools==0.6.4
idna==2.8
jose==1.0.0
lockfile==0.12.2
Mako==1.3.9
MarkupSafe==3.0.2
passlib==1.7.4
psycopg==3.2.6
pyasn1==0.4.8
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==1.7.1
python-daemon==2.2.4
python-dotenv==1.0.1
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.22.0
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.39
starlette==0.46.1
typing_extensions==4.12.2
urllib3==1.25.11
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
websockets==15.0.1
